forum-service/requirements.txt
asgiref             3.8.1
Django              5.2.1
djangorestframework 3.16.0
pip                 25.1.1
psycopg2-binary     2.9.10
python-decouple     3.8
sqlparse            0.5.3
